Extract country selector to its own component
