+++ src/components.d.ts
@@ -11,6 +11,7 @@ declare module 'vue' {
     AdminLevelToggle: typeof import('./components/mapSettingsMenu/AdminLevelToggle.vue')['default']
     Choropleth: typeof import('./components/Choropleth.vue')['default']
     ColorScaleIcon: typeof import('./components/mapSettingsMenu/ColorScaleIcon.vue')['default']
+    CountrySelector: typeof import('./components/mapSettingsMenu/CountrySelector.vue')['default']
     ExcelDownloadButton: typeof import('./components/ExcelDownloadButton.vue')['default']
     HelpAlert: typeof import('./components/HelpAlert.vue')['default']
     IndicatorMenu: typeof import('./components/indicatorMenu/IndicatorMenu.vue')['default']
